Rework ammo pickups
Ammo pickups are now based on weapon damage instead of total ammo and is calculated automatically for every weapon.
In addition any fractional part of an ammo pickup roll will be carried over to the next roll so pickup is more consistent over time and less RNG.
With the above changes, LMGs have seen a slight reduction in their pickup values but are still manageable if not spammed excessively.
Pistols and high damage shotguns will see a small improve in their average pickup values.